+++ DirectX11/HackerDevice.cpp
@@ -1636,7 +1636,7 @@ STDMETHODIMP HackerDevice::CreateUnorderedAccessView(THIS_
 				}
 			}
 
-			if (override_num_elements != -1 && pDesc->Buffer.NumElements < override_num_elements) {
+			if (override_num_elements != -1 && pDesc->Buffer.NumElements < (UINT)override_num_elements) {
 				D3D11_UNORDERED_ACCESS_VIEW_DESC pNewDesc = *pDesc;
 				pNewDesc.Buffer.NumElements = override_num_elements;
 				//LogOverlayW(LOG_INFO, L"UAV resized: %d->%d\n", pDesc->Buffer.NumElements, override_num_elements);
@@ -2075,7 +2075,7 @@ static void override_resource_desc_common_2d_3d(DescType *desc, TextureOverride 
 
 static void override_resource_desc(D3D11_BUFFER_DESC *desc, TextureOverride *textureOverride) {
 	if (textureOverride->override_byte_width != -1) {
-		if (desc->ByteWidth < textureOverride->override_byte_width) {
+		if (desc->ByteWidth < (UINT)textureOverride->override_byte_width) {
 			LogInfo("  resizing buffer: %d->%d\n", desc->ByteWidth, textureOverride->override_byte_width);
 			//LogOverlayW(LOG_WARNING, L"Buffer resized: %d->%d\n - [%s]\n", desc->ByteWidth, textureOverride->override_byte_width, textureOverride->ini_section.c_str());
 			desc->ByteWidth = textureOverride->override_byte_width;
